requirements.txt
fastapi
uvicorn
numpy
python-telegram-bot
